Refactor methods and classes.
